refactor(micromouse): パラメータ初期化抽出・リーク修正・include整理(ロジック不変)
run_micromouse の内部改善(挙動不変)。
- パラメータ初期化(gain/閾値/スラローム/ゴール ~135行)を set_default_params() へ抽出。
  ブロックはバイト等価で移動(値・順序・代入先すべて不変)、本体の見通しを改善。
- set_interface() に `if (!ui.empty()) return;` を追加。mode_select 毎に UI を16個
  push_back し続けるリークを解消(ui[mode] は常に先頭16個を指すため選択タスク・実行挙動は不変)。
- include を "include/micromouse.hpp" → "micromouse.hpp"(INCLUDE_DIRS 経由)へ整理。

要ビルド検証。

+++ components/mouse/motion/motion.cpp
@@ -536,7 +536,7 @@ void ::Motion::stop2()
     // 壁補正(コメントアウト中の前壁補正ロジック)で使用する変数。再有効化に備え温存。
     // 注: hosei_dist は uint8_t に 0.050 を代入しており 0 に切り捨てられる(再有効化時は型を float へ要修正)。
     [[maybe_unused]] bool hosei_flag = NOT_YET;
-    [[maybe_unused]] uint8_t hosei_dist = 0.050;
+    [[maybe_unused]] float hosei_dist = 0.050;
 
     // 角度制御の準備（壁制御がFALSEなので角度制御を有効化）
     val->start_angle = val->current.rad;
